--- a/Engine/HistoricalData/SynchronizingHistoryProvider.cs
+++ b/Engine/HistoricalData/SynchronizingHistoryProvider.cs
@@ -15,6 +15,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading;
 using NodaTime;
 using QuantConnect.Data;
@@ -53,9 +54,14 @@ namespace QuantConnect.Lean.Engine.HistoricalData
             {
                 var earlyBirdTicks = long.MaxValue;
                 var data = new List<DataFeedPacket>();
-                foreach (var subscription in subscriptions)
+                foreach (var subscription in subscriptions.Where(subscription => !subscription.EndOfStream))
                 {
-                    if (subscription.EndOfStream) continue;
+                    if (subscription.Current == null && !subscription.MoveNext())
+                    {
+                        // initial pump. We do it here and not when creating the subscriptions so
+                        // that parallel workers can all start as fast as possible
+                        continue;
+                    }
 
                     var packet = new DataFeedPacket(subscription.Security, subscription.Configuration);
 
@@ -70,7 +76,7 @@ namespace QuantConnect.Lean.Engine.HistoricalData
                     }
                     // only add if we have data
                     if (packet.Count != 0) data.Add(packet);
-                    // udate our early bird ticks (next frontier time)
+                    // update our early bird ticks (next frontier time)
                     if (subscription.Current != null)
                     {
                         // take the earliest between the next piece of data or the next tz discontinuity
@@ -78,14 +84,14 @@ namespace QuantConnect.Lean.Engine.HistoricalData
                     }
                 }
 
-                // end of subscriptions
-                if (earlyBirdTicks == long.MaxValue) break;
-
                 if (data.Count != 0)
                 {
                     // reuse the slice construction code from TimeSlice.Create
                     yield return timeSliceFactory.Create(frontier, data, SecurityChanges.None, new Dictionary<Universe, BaseDataCollection>()).Slice;
                 }
+
+                // end of subscriptions, after we emit, else we might drop a data point
+                if (earlyBirdTicks == long.MaxValue) break;
 
                 frontier = new DateTime(Math.Max(earlyBirdTicks, frontier.Ticks), DateTimeKind.Utc);
             }
@@ -102,10 +108,6 @@ namespace QuantConnect.Lean.Engine.HistoricalData
         /// </summary>
         protected Subscription CreateSubscription(HistoryRequest request, IEnumerable<BaseData> history)
         {
-            // data reader expects these values in local times
-            var start = request.StartTimeUtc.ConvertFromUtc(request.ExchangeHours.TimeZone);
-            var end = request.EndTimeUtc.ConvertFromUtc(request.ExchangeHours.TimeZone);
-
             var config = new SubscriptionDataConfig(request.DataType,
                 request.Symbol,
                 request.Resolution,
@@ -135,6 +137,10 @@ namespace QuantConnect.Lean.Engine.HistoricalData
             // optionally apply fill forward behavior
             if (request.FillForwardResolution.HasValue)
             {
+                // FillForwardEnumerator expects these values in local times
+                var start = request.StartTimeUtc.ConvertFromUtc(request.ExchangeHours.TimeZone);
+                var end = request.EndTimeUtc.ConvertFromUtc(request.ExchangeHours.TimeZone);
+
                 // copy forward Bid/Ask bars for QuoteBars
                 if (request.DataType == typeof(QuoteBar))
                 {
@@ -142,13 +148,13 @@ namespace QuantConnect.Lean.Engine.HistoricalData
                 }
 
                 var readOnlyRef = Ref.CreateReadOnly(() => request.FillForwardResolution.Value.ToTimeSpan());
-                reader = new FillForwardEnumerator(reader, security.Exchange, readOnlyRef, security.IsExtendedMarketHours, end, config.Increment, config.DataTimeZone, start);
+                reader = new FillForwardEnumerator(reader, security.Exchange, readOnlyRef, request.IncludeExtendedMarketHours, end, config.Increment, config.DataTimeZone, start);
             }
 
-            var timeZoneOffsetProvider = new TimeZoneOffsetProvider(security.Exchange.TimeZone, start, end);
+            var timeZoneOffsetProvider = new TimeZoneOffsetProvider(security.Exchange.TimeZone, request.StartTimeUtc, request.EndTimeUtc);
             var subscriptionDataEnumerator = new SubscriptionDataEnumerator(config, security.Exchange.Hours, timeZoneOffsetProvider, reader);
 
-            var subscriptionRequest = new SubscriptionRequest(false, null, security, config, start, end);
+            var subscriptionRequest = new SubscriptionRequest(false, null, security, config, request.StartTimeUtc, request.EndTimeUtc);
             return new Subscription(subscriptionRequest, subscriptionDataEnumerator, timeZoneOffsetProvider);
         }
     }